Split out evaluators into separate scripts
