+++ Assets/Scripts/Incidents/CriteriaEvaluators/FloatEvaluator.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace Game.Incidents
+{
+	public class FloatEvaluator : CriteriaEvaluator<float>
+    {
+        public FloatEvaluator() : base() { }
+        public FloatEvaluator(string propertyName) : base(propertyName) { }
+        public FloatEvaluator(string propertyName, float value) : base(propertyName, value) { }
+        public FloatEvaluator(string propertyName, Type contextType) : base(propertyName, contextType) { }
+
+        override protected void Setup()
+        {
+            Comparators = ExpressionHelpers.FloatComparators;
+            Operators = ExpressionHelpers.FloatOperators;
+        }
+    }
+}
